fix 32, 64-bit issue on Windows

+++ transformer/Beam.py
@@ -87,7 +87,7 @@ class Beam(object):
             _, keys = self.sort_scores()
             hyps = [self.get_hypothesis(k) for k in keys]
             hyps = [[Constants.BOS] + h for h in hyps]
-            dec_seq = torch.from_numpy(np.array(hyps))
+            dec_seq = torch.from_numpy(np.array(hyps, dtype=np.int64))
 
         return dec_seq
